refactor(frontend): rename Settings component
